Add power management (Add PM5CTL0 set to high to lock pin I/O conf)
While looking through the disassebly I saw a reference to PM5CTL0 which is a
control bit for Power Mode 5; essentially if set to high this locks
GPIO configuration, which prevents it from being reset and therefor avoiding
the P4DIR & P4OUT values being set.

I'm not sure if this is 100% correct, but it works!